streamlit_portal/requirements.txt
streamlit
azure-cosmos
python-dotenv
